Fix explicit package version install when not on the latest version.
The server package was getting installed first which installed the latest
version of the client package as a dependency if nothing else was installed.
This then caused issues when we tried to install an explicit version of the
client package afterwards (since the newer version was already installed). The
fix is to simply swap the install order so the client package dependency gets
installed first with an explicit version.